move CvxCompress compression option to an extension
