+++ Project.toml
@@ -1,6 +1,6 @@
 name = "CloudSeis"
 uuid = "a8ea0714-0c8b-51e5-9b11-67a2d82946d2"
-version = "1.11.0"
+version = "1.12.0"
 
 [deps]
 AbstractStorage = "14dbef02-f468-5f15-853e-5ec8dee7b899"
@@ -13,6 +13,15 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 TeaSeis = "db5f7d96-a200-5343-9fc6-a259b42289b2"
 UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
 ZfpCompression = "43441a71-1662-41c6-b8ea-40ed1525242b"
+
+[weakdeps]
+CvxCompress = "3e489a4b-a92a-5e1b-a7bf-ed666eae205e"
+
+[extras]
+CvxCompress = "3e489a4b-a92a-5e1b-a7bf-ed666eae205e"
+
+[extensions]
+CvxCompressExt = "CvxCompress"
 
 [compat]
 AbstractStorage = "^1.1"
