+++ test/Project.toml
@@ -1,6 +1,7 @@
 [deps]
 AzSessions = "f239b30d-ae6b-58be-a2d5-7e9f30e280a9"
 AzStorage = "c6697862-1611-5eae-9ef8-48803c85c8d6"
+CvxCompress = "3e489a4b-a92a-5e1b-a7bf-ed666eae205e"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
 FolderStorage = "37e990fb-a500-53d7-846b-4c8aba66999e"
